+++ scripts/macros/2024/spells/spiritualWeapon.js
@@ -1,0 +1,6 @@
+import {spiritualWeapon as spiritualWeaponLegacy} from '../../../legacyMacros.js';
+export let spiritualWeapon = {
+    ...spiritualWeaponLegacy,
+    version: '1.2.24',
+    rules: 'modern'
+};